+++ src/sc_keeper/parameters.py
@@ -22,6 +22,7 @@ from .references import (
     ComplianceFrameworks,
     Countries,
     CpuFamilies,
+    CpuFlags,
     CpuL1CacheSnapPoints,
     CpuL1CacheTotalSnapPoints,
     CpuL2CacheSnapPoints,
@@ -133,6 +134,18 @@ cpu_family = Annotated[
         json_schema_extra={
             "category_id": CommonFilterCategory.PROCESSOR,
             "enum": [e.value for e in CpuFamilies],
+        },
+    ),
+]
+
+cpu_flags = Annotated[
+    Optional[List[CpuFlags]],
+    Query(
+        title="CPU flags",
+        description="Required CPU flags.",
+        json_schema_extra={
+            "category_id": CommonFilterCategory.PROCESSOR,
+            "enum": [e.value for e in CpuFlags],
         },
     ),
 ]
@@ -268,7 +281,7 @@ network_speed_baseline_min = Annotated[
         title="Required baseline network speed",
         description="Required baseline network speed in Gbps.",
         json_schema_extra={
-            "category_id": ServerFilterCategory.TRAFFIC,
+            "category_id": CommonFilterCategory.TRAFFIC,
             "enum": [e.value for e in NetworkSpeedSnapPoints],
             "unit": "Gbps",
         },
@@ -281,7 +294,7 @@ network_speed_max_min = Annotated[
         title="Required maximum network speed",
         description="Required maximum network speed in Gbps.",
         json_schema_extra={
-            "category_id": ServerFilterCategory.TRAFFIC,
+            "category_id": CommonFilterCategory.TRAFFIC,
             "enum": [e.value for e in NetworkSpeedSnapPoints],
             "unit": "Gbps",
         },
@@ -414,6 +427,18 @@ hw_virt = Annotated[
     ),
 ]
 
+cpu_hyperthreading = Annotated[
+    Optional[bool],
+    Query(
+        title="Hyperthreading",
+        description=(
+            'Whether CPU hyperthreading is enabled or not. Calculated based on the "ht" CPU flag, '
+            "with a fallback to comparing the number of vCPUs with physical CPU cores."
+        ),
+        json_schema_extra={"category_id": CommonFilterCategory.PROCESSOR},
+    ),
+]
+
 storage_size = Annotated[
     Optional[float],
     Query(
@@ -446,7 +471,7 @@ network_storage_speed_baseline_min = Annotated[
         title="Required baseline network storage speed",
         description="Required baseline network storage speed in Gbps.",
         json_schema_extra={
-            "category_id": ServerFilterCategory.TRAFFIC,
+            "category_id": CommonFilterCategory.TRAFFIC,
             "enum": [e.value for e in NetworkStorageSpeedSnapPoints],
             "unit": "Gbps",
         },
@@ -459,7 +484,7 @@ network_storage_speed_max_min = Annotated[
         title="Required maximum network storage speed",
         description="Required maximum network storage speed in Gbps.",
         json_schema_extra={
-            "category_id": ServerFilterCategory.TRAFFIC,
+            "category_id": CommonFilterCategory.TRAFFIC,
             "enum": [e.value for e in NetworkStorageSpeedSnapPoints],
             "unit": "Gbps",
         },
@@ -475,7 +500,7 @@ monthly_inbound_traffic = Annotated[
             "The cheapest available inbound traffic price for the vendor is used."
         ),
         json_schema_extra={
-            "category_id": ServerFilterCategory.TRAFFIC,
+            "category_id": CommonFilterCategory.TRAFFIC,
             "unit": "GB",
             "step": 1,
         },
@@ -491,7 +516,7 @@ monthly_outbound_traffic = Annotated[
             "The cheapest available outbound traffic price for the vendor is used."
         ),
         json_schema_extra={
-            "category_id": ServerFilterCategory.TRAFFIC,
+            "category_id": CommonFilterCategory.TRAFFIC,
             "unit": "GB",
             "step": 1,
         },
@@ -537,7 +562,7 @@ direction = Annotated[
         title="Direction",
         description="Direction of the Internet traffic.",
         json_schema_extra={
-            "category_id": ServerFilterCategory.TRAFFIC,
+            "category_id": CommonFilterCategory.TRAFFIC,
             "enum": [e.value for e in TrafficDirection],
         },
     ),
@@ -549,7 +574,7 @@ monthly_traffic = Annotated[
         title="Monthly overall traffic",
         description="Overall amount of monthly traffic (GBs).",
         json_schema_extra={
-            "category_id": ServerFilterCategory.TRAFFIC,
+            "category_id": CommonFilterCategory.TRAFFIC,
             "unit": "GB",
             "step": 1,
         },
@@ -651,7 +676,7 @@ benchmark_score_stressng_cpu_min = Annotated[
         title="Required SCore",
         description="Required stress-ng div16 CPU workload score.",
         json_schema_extra={
-            "category_id": ServerFilterCategory.PERFORMANCE,
+            "category_id": CommonFilterCategory.PERFORMANCE,
         },
     ),
 ]
@@ -663,7 +688,7 @@ benchmark_score_per_price_stressng_cpu_min = Annotated[
         title="Required $Core",
         description="Required stress-ng div16 CPU workload score per USD/hr (using the best ondemand or spot price of all zones).",
         json_schema_extra={
-            "category_id": ServerFilterCategory.PERFORMANCE,
+            "category_id": CommonFilterCategory.PERFORMANCE,
             "unit": "/USD",
         },
     ),
@@ -693,7 +718,7 @@ benchmark_score_min = Annotated[
         title="Required benchmark score",
         description="Required value of the selected benchmark score.",
         json_schema_extra={
-            "category_id": ServerFilterCategory.PERFORMANCE,
+            "category_id": CommonFilterCategory.PERFORMANCE,
         },
     ),
 ]
@@ -705,7 +730,7 @@ benchmark_score_per_price_min = Annotated[
         title="Required benchmark score/price",
         description="Required value of the selected benchmark score per USD/hr (using the best ondemand or spot price of all zones).",
         json_schema_extra={
-            "category_id": ServerFilterCategory.PERFORMANCE,
+            "category_id": CommonFilterCategory.PERFORMANCE,
             "unit": "/USD",
         },
     ),
@@ -1066,6 +1091,31 @@ database_sla_min = Annotated[
         json_schema_extra={
             "category_id": DatabaseFilterCategory.FEATURES,
             "step": 0.01,
+        },
+    ),
+]
+
+database_benchmark_score_min = Annotated[
+    Optional[float],
+    Query(
+        title="Required benchmark score",
+        description=("Required value of the selected benchmark score."),
+        json_schema_extra={
+            "category_id": CommonFilterCategory.PERFORMANCE,
+        },
+    ),
+]
+
+database_benchmark_score_per_price_min = Annotated[
+    Optional[float],
+    Query(
+        title="Required benchmark score/price",
+        description=(
+            "Required value of the selected benchmark score per USD/hr (using the best on-demand price)."
+        ),
+        json_schema_extra={
+            "category_id": CommonFilterCategory.PERFORMANCE,
+            "unit": "/USD",
         },
     ),
 ]